+++ tests/fakeredis/test/test_hypothesis/base.py
@@ -2,7 +2,7 @@ import functools
 import math
 import operator
 import sys
-from typing import Any, Tuple, Union
+from typing import Any
 
 import fakeredis
 import hypothesis
@@ -10,9 +10,10 @@ import hypothesis.stateful
 import hypothesis.strategies as st
 import pytest
 import redis
-from fakeredis._server import _create_version
 from hypothesis.stateful import rule, initialize, precondition
 from hypothesis.strategies import SearchStrategy
+
+from ._server_info import redis_ver
 
 self_strategy = st.runner()
 
@@ -29,29 +30,6 @@ def sample_attr(draw, name):
     return values[position]
 
 
-def server_info() -> Tuple[str, Union[None, Tuple[int, ...]]]:
-    """Returns server's version or None if server is not running"""
-    client = None
-    try:
-        client = redis.Redis("localhost", port=6380, db=2)
-        client_info = client.info()
-        server_type = "dragonfly" if "dragonfly_version" in client_info else "redis"
-        server_version = (
-            client_info["redis_version"] if server_type != "dragonfly" else (7, 0)
-        )
-        server_version = _create_version(server_version) or (7,)
-        return server_type, server_version
-    except redis.ConnectionError as e:
-        print(e)
-        pytest.exit("Redis is not running")
-        return "redis", (6,)
-    finally:
-        if hasattr(client, "close"):
-            client.close()  # Absent in older versions of redis-py
-
-
-server_type, redis_ver = server_info()
-
 keys = sample_attr("keys")
 fields = sample_attr("fields")
 values = sample_attr("values")
@@ -59,7 +37,6 @@ scores = sample_attr("scores")
 
 ints = st.integers(min_value=MIN_INT, max_value=MAX_INT)
 int_as_bytes = st.builds(lambda x: str(default_normalize(x)).encode(), ints)
-optional_bitcount_range = st.just(()) | st.tuples(int_as_bytes, int_as_bytes)
 float_as_bytes = st.builds(
     lambda x: repr(default_normalize(x)).encode(), st.floats(width=32)
 )
@@ -267,14 +244,6 @@ zset_no_score_commands = (  # TODO: test incr
     | commands(st.just("zremrangebylex"), keys, string_tests, string_tests)
 )
 
-bad_commands = (
-    # redis-py splits the command on spaces, and hangs if that ends up
-    # being an empty list
-    commands(
-        st.text().filter(lambda x: bool(x.split())), st.lists(st.binary() | st.text())
-    )
-)
-
 attrs = st.fixed_dictionaries(
     {
         "keys": st.lists(st.binary(), min_size=2, max_size=5, unique=True),
@@ -429,269 +398,3 @@ class BaseTest:
         )
         hypothesis.settings.load_profile("debug")
         hypothesis.stateful.run_state_machine_as_test(Machine)
-
-
-class TestConnection(BaseTest):
-    # TODO: tests for select
-    connection_commands = (
-        commands(st.just("echo"), values)
-        | commands(st.just("ping"), st.lists(values, max_size=2))
-        # | commands(st.just("swapdb"), dbnums, dbnums)
-    )
-    command_strategy = connection_commands | common_commands
-
-
-class TestString(BaseTest):
-    string_commands = (
-        commands(st.just("append"), keys, values)
-        | commands(st.just("bitcount"), keys, optional_bitcount_range)
-        | commands(st.sampled_from(["incr", "decr"]), keys)
-        | commands(st.sampled_from(["incrby", "decrby"]), keys, values)
-        | commands(st.just("get"), keys)
-        | commands(st.just("getbit"), keys, counts)
-        | commands(
-            st.just("setbit"),
-            keys,
-            counts,
-            st.integers(min_value=0, max_value=1) | ints,
-        )
-        | commands(st.sampled_from(["substr", "getrange"]), keys, counts, counts)
-        | commands(st.just("getset"), keys, values)
-        | commands(st.just("mget"), st.lists(keys))
-        | commands(
-            st.sampled_from(["mset", "msetnx"]), st.lists(st.tuples(keys, values))
-        )
-        | commands(
-            st.just("set"),
-            keys,
-            values,
-            *zero_or_more("nx", "xx", "keepttl"),
-        )
-        | commands(st.just("setex"), keys, expires_seconds, values)
-        | commands(st.just("psetex"), keys, expires_ms, values)
-        | commands(st.just("setnx"), keys, values)
-        | commands(st.just("setrange"), keys, counts, values)
-        | commands(st.just("strlen"), keys)
-    )
-    create_command_strategy = commands(st.just("set"), keys, values)
-    command_strategy = string_commands | common_commands
-
-
-class TestHash(BaseTest):
-    hash_commands = (
-        commands(st.just("hset"), keys, st.lists(st.tuples(fields, values)))
-        | commands(st.just("hdel"), keys, st.lists(fields))
-        | commands(st.just("hexists"), keys, fields)
-        | commands(st.just("hget"), keys, fields)
-        | commands(st.sampled_from(["hgetall", "hkeys", "hvals"]), keys)
-        | commands(st.just("hincrby"), keys, fields, ints)
-        | commands(st.just("hlen"), keys)
-        | commands(st.just("hmget"), keys, st.lists(fields))
-        | commands(st.just("hset"), keys, st.lists(st.tuples(fields, values)))
-        | commands(st.just("hsetnx"), keys, fields, values)
-        | commands(st.just("hstrlen"), keys, fields)
-        | commands(
-            st.just("hpersist"),
-            st.just("fields"),
-            st.just(2),
-            st.lists(fields, min_size=2, max_size=2),
-        )
-        | commands(
-            st.just("hexpire"),
-            keys,
-            expires_seconds,
-            # TODO: Dragonfly does not support the following arguments
-            # *zero_or_more("nx", "xx", "gt", "lt"),
-            st.just("fields"),
-            st.just(2),
-            st.lists(fields, min_size=2, max_size=2),
-        )
-    )
-    create_command_strategy = commands(
-        st.just("hset"), keys, st.lists(st.tuples(fields, values), min_size=1)
-    )
-    command_strategy = hash_commands | common_commands
-
-
-class TestList(BaseTest):
-    # TODO: blocking commands
-    list_commands = (
-        commands(st.just("lindex"), keys, counts)
-        | commands(
-            st.just("linsert"),
-            keys,
-            st.sampled_from(["before", "after", "BEFORE", "AFTER"]) | st.binary(),
-            values,
-            values,
-        )
-        | commands(st.just("llen"), keys)
-        | commands(
-            st.sampled_from(["lpop", "rpop"]),
-            keys,
-            st.just(None) | st.just([]) | ints,
-        )
-        | commands(
-            st.sampled_from(["lpush", "lpushx", "rpush", "rpushx"]),
-            keys,
-            st.lists(values),
-        )
-        | commands(st.just("lrange"), keys, counts, counts)
-        | commands(st.just("lrem"), keys, counts, values)
-        | commands(st.just("lset"), keys, counts, values)
-        | commands(st.just("ltrim"), keys, counts, counts)
-        | commands(st.just("rpoplpush"), keys, keys)
-    )
-    create_command_strategy = commands(
-        st.just("rpush"), keys, st.lists(values, min_size=1)
-    )
-    command_strategy = list_commands | common_commands
-
-
-class TestSet(BaseTest):
-    set_commands = (
-        commands(st.just("sadd"), keys, st.lists(fields))
-        | commands(st.just("scard"), keys)
-        | commands(st.sampled_from(["sdiff", "sinter", "sunion"]), st.lists(keys))
-        | commands(
-            st.sampled_from(["sdiffstore", "sinterstore", "sunionstore"]),
-            keys,
-            st.lists(keys),
-        )
-        | commands(st.just("sismember"), keys, fields)
-        | commands(st.just("smembers"), keys)
-        | commands(st.just("smove"), keys, keys, fields)
-        | commands(st.just("srem"), keys, st.lists(fields))
-    )
-    # TODO:
-    # - find a way to test srandmember, spop which are random
-    # - sscan
-    create_command_strategy = commands(
-        st.just("sadd"), keys, st.lists(fields, min_size=1)
-    )
-    command_strategy = set_commands | common_commands
-
-
-class TestZSet(BaseTest):
-    zset_commands = (
-        commands(
-            st.just("zadd"),
-            keys,
-            *zero_or_more("nx", "xx", "ch", "incr"),
-            st.lists(st.tuples(scores, fields)),
-        )
-        | commands(st.just("zcard"), keys)
-        | commands(st.just("zcount"), keys, score_tests, score_tests)
-        | commands(st.just("zincrby"), keys, scores, fields)
-        | commands(
-            st.sampled_from(["zrange", "zrevrange"]),
-            keys,
-            counts,
-            counts,
-            optional("withscores"),
-        )
-        | commands(
-            st.sampled_from(["zrangebyscore", "zrevrangebyscore"]),
-            keys,
-            score_tests,
-            score_tests,
-            limits,
-            optional("withscores"),
-        )
-        | commands(st.sampled_from(["zrank", "zrevrank"]), keys, fields)
-        | commands(st.just("zrem"), keys, st.lists(fields))
-        | commands(st.just("zremrangebyrank"), keys, counts, counts)
-        | commands(st.just("zremrangebyscore"), keys, score_tests, score_tests)
-        | commands(st.just("zscore"), keys, fields)
-        | st.builds(
-            build_zstore,
-            command=st.sampled_from(["zunionstore", "zinterstore"]),
-            dest=keys,
-            sources=st.lists(st.tuples(keys, float_as_bytes)),
-            weights=st.booleans(),
-            aggregate=st.sampled_from([None, "sum", "min", "max"]),
-        )
-    )
-    # TODO: zscan, zpopmin/zpopmax, bzpopmin/bzpopmax, probably more
-    create_command_strategy = commands(
-        st.just("zadd"), keys, st.lists(st.tuples(scores, fields), min_size=1)
-    )
-    command_strategy = zset_commands | common_commands
-
-
-class TestZSetNoScores(BaseTest):
-    create_command_strategy = zset_no_score_create_commands
-    command_strategy = zset_no_score_commands | common_commands
-
-
-class TestTransaction(BaseTest):
-    transaction_commands = (
-        commands(st.sampled_from(["multi", "discard", "exec", "unwatch"]))
-        | commands(st.just("watch"), keys)
-        | commands(st.just("append"), keys, values)
-        | commands(st.just("bitcount"), keys)
-        | commands(st.just("bitcount"), keys, values, values)
-        | commands(st.sampled_from(["incr", "decr"]), keys)
-        | commands(st.sampled_from(["incrby", "decrby"]), keys, values)
-        | commands(st.just("get"), keys)
-        | commands(st.just("getbit"), keys, counts)
-        | commands(
-            st.just("setbit"),
-            keys,
-            counts,
-            st.integers(min_value=0, max_value=1) | ints,
-        )
-        | commands(st.sampled_from(["substr", "getrange"]), keys, counts, counts)
-        | commands(st.just("getset"), keys, values)
-        | commands(st.just("mget"), st.lists(keys))
-        | commands(
-            st.sampled_from(["mset", "msetnx"]), st.lists(st.tuples(keys, values))
-        )
-        | commands(
-            st.just("set"),
-            keys,
-            values,
-            *zero_or_more("nx", "xx", "keepttl"),
-        )
-        | commands(st.just("setex"), keys, expires_seconds, values)
-        | commands(st.just("psetex"), keys, expires_ms, values)
-        | commands(st.just("setnx"), keys, values)
-        | commands(st.just("setrange"), keys, counts, values)
-        | commands(st.just("strlen"), keys)
-    )
-    create_command_strategy = TestString.create_command_strategy
-    command_strategy = transaction_commands | common_commands
-
-
-class TestServer(BaseTest):
-    # TODO: real redis raises an error if there is a save already in progress.
-    #  Find a better way to test this. commands(st.just('bgsave'))
-    server_commands = (
-        commands(st.just("dbsize"))
-        | commands(st.sampled_from(["flushdb", "flushall"]))
-        # TODO: result is non-deterministic
-        # | commands(st.just('lastsave'))
-        | commands(st.just("save"))
-    )
-    create_command_strategy = TestString.create_command_strategy
-    command_strategy = server_commands | TestString.string_commands | common_commands
-
-
-class TestJoint(BaseTest):
-    create_command_strategy = (
-        TestString.create_command_strategy
-        | TestHash.create_command_strategy
-        | TestList.create_command_strategy
-        | TestSet.create_command_strategy
-        | TestZSet.create_command_strategy
-    )
-    command_strategy = (
-        TestServer.server_commands
-        | TestConnection.connection_commands
-        | TestString.string_commands
-        | TestHash.hash_commands
-        | TestList.list_commands
-        | TestSet.set_commands
-        | TestZSet.zset_commands
-        | common_commands
-        | bad_commands
-    )
